Fix: external storage problem NavBar applications
If the user binds an app which is located on the
external sdcard and is rebooting the button icon
will be left blank because the navbar is already
constructed before the sdcard is mounted. To avoid
this we listen to the app available intents.
If it is not available it shows the
none graphics. When available it shows the
correct app icon. So that the user can clearly see
this app is accessable or not. As well when the
user goes into mass storage mode or does
unmount the sdcard or whatever case could
happen that the application is not there.

+ a small optimize

cheers

Change-Id: Ic6a691d8b3af9ddb3e6008d64e6ab35cbabc12fb