Frameworks: fcharge for all (1/2)
Thanks roman for the basic idea to use an overlay
to set the fast charge path so each device can set
the propper path to the custom kernel.
As well this commit pays attention if the user
flashed a kernel which does not support fastcharge
or is using another path. Option for tile in settings will be removed
as well.

not 100% happy with the whole fast charge implementation...but I am lazy :p
...whatever maybe I will recode that all on a later point. It
is working as intended like it is.

Change-Id: I0cb830c1ecfebe479a2e10a31c7c84967e2e2030
Signed-off-by: scotthartbti <scott@ubuntu.com>

Conflicts:
	core/res/res/values/config.xml