Framework: Always show settings button option for notification drawer (1/2)
This also fixes a bug where the settings button was not showing the proper pressed drawable
and fixes the transparency on the ic_notify_settings_normal drawable to match the rest of the
notification drawer assets.

Also removes longpress PW toggle setting since it now defaults to this.

Change-Id: I1b7624fd3278504a07c79daab00babcb5f7b37d1
Signed-off-by: scotthartbti <scott.hart.bti@gmail.com>

Conflicts:
	core/java/android/provider/Settings.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarView.java

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java