Disable Google's home-key handling code
This might be temporary. Upstream now includes multi-behavior key
handling for HOME, which conflicts with our own code when running
in parallel, resulting in funky behaviors. Since the CM implementation
is more flexible, let's keep that for the time being and disable
Google's; It might be a good idea to rewrite our version so that it's
an extension of theirs, though.

Change-Id: I5d55301772607315b410880c1194a0076e21cb80
Signed-off-by: scotthartbti <scott@ubuntu.com>

Conflicts:
	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java