Remove redundant click action from date & time container layout
Unlike AOSP, we  provide separate actions for both the clock
and date & time views in the expanded notification area. AOSP has a
single action for both views.

This commit removes the redundant AOSP click action completely. In
addition, the clock and date views are modified to occupy the entire
height of the layout to increase the clickable area.

Change-Id: I78f4353c9ae4646f14124a38ef94e44ab920a067
Signed-off-by: scott <scott@ubuntu.com>

Conflicts:
	packages/SystemUI/res/layout/status_bar_expanded_header.xml