fb: Fix wrong IME state handlings on keyboard switches
this change introduced some slim keyboard features

https://github.com/SlimRoms/frameworks_base/commit/d51b14e667fe0dfed02c5080fbea97c349a12889

We assumed that calling updateFromSettingsLocked(true) is in all value changes
a possible way to simplify the process. In reality this fails when the user
switches eg between 2 IME keyboards several times due that we call every
time

                        mIPackageManager.setApplicationEnabledSetting(imm.getPackageName(),
                                PackageManager.COMPONENT_ENABLED_STATE_DEFAULT,
                                PackageManager.DONT_KILL_APP, mSettings.getCurrentUserId(),
                                mContext.getBasePackageName());

which results in wrong mCurMethodId which ends in method calling which should not be called
and result in an NPE and hotreboot due of dieing service.

To fix it go only for showinputmethod feature the old way and restore for all other
value changes AOSP behaviour and check specifically if enabledChanged is true or not
based on the getEnabledInputMethodsStr() method during we lock on mMethodMap

Thanks for the detailed report to Natakranta

Change-Id: I2931ca5a65c56999ba9ce7a7f4f7e71437a4bec8
Signed-off-by: scott <scott@ubuntu>

Conflicts:
	services/java/com/android/server/InputMethodManagerService.java