HOVER: touch to hide (1/2)
- Allow users to configure touching outside to instant hide HOVER notifications

PS: This commit brings back the -> private static final int LONG_FADE_OUT_DELAY along with a new simple method public void startFiveSecHideCountdown(). While i was working on this feature i've noted a bad behavior in hover notification timeout implementation thats was if we just set, for example 15 seconds of timeout, the touch outside would read this as hide timeout in expanded hover notifications. 3 or 5 seconds are good but anything above that is too much. Back to normal with the new method fixing 5 seconds as the maximum
Signed-off-by: scott <scott@ubuntu>

Conflicts:
	core/java/android/provider/Settings.java
	packages/SystemUI/src/com/android/systemui/statusbar/notification/Hover.java

Change-Id: I55df8d73c2270c55e7eeea0c7c77bc98c761429d

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/notification/Hover.java