Profiles: Disable lock screen after boot
This brings it inline with stock behavior and also fixes a couple of issues.
If lock screen was disabled via profiles, on boot you would end up with no home or recents buttons
and if you had a simpin set you would be faced with a black screen with no way to access the phone unless you
popped the sim out.

Ref: CYNGNOS-1930

Change-Id: I9f1f9f467647e31525a1179dc1a0ce7b9dc3285f