Statusbar: Gracefully adapt 3Minit Battery hook for AOSP
3Minit Battery is a free or paid third party statusbar
battery icon solution. This commit builds upon the
initial SystemUI hook source release by Gary Harrington.

We are able to significantly optimize the original code
because we are not constrained by the "one-size-fits-all"
conditions the original code must meet. At the same time,
we fully maintain the integrity of the UX intended by
the original code.

Of note is the introduction of a controller class. This
class takes over settings and state management as well
as handles receiver events. The controller class also
monitors package changes and updates feature visibility
accordingly if a needed resource is added or removed.

Upon initialization, if no feature resources are
detected, all resources are freed.

** Changes for N **

There is no longer a batter view in the statusbar header
so don't pass as constructor arg and adapt in controller
class.

Change-Id: I0254bfa0becfbf6e83e9d0547abda8a77047ab1e
Signed-off-by: Varun Date <date.varun123@gmail.com>
Signed-off-by: scotthartbti <scott.hart.bti@gmail.com>