SystemUI: improve layout of qs detail panel
Don't try to fit the detail views, but allow the user to scroll freely
while detail view is showing and we are in landscape. The user can still
fling up and down and the velocity flings are respected in the detail
scroll lock mode.

Hitting done will always return the user
to the QS panel, or they can close the detail view manually.

Ticket: FEIJ-201

Change-Id: Id8b16a682ba5cebdc57c0455d81aa38fd5e90cbd
Signed-off-by: Roman Birg <roman@cyngn.com>
Signed-off-by: scott <scott.hart.bti@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelView.java