Remove default Gapps permissions.  This is not necessary.
I noticed this with Android 6.0 as well as now with Android 7.0.. Not
sure how this all started, but setting default Gapps permissions is
not necessary.  This can be completely removed and Gapps setup and the
Gapps applications will perform as expected without issue.

More information found here: http://review.cyanogenmod.org/#/c/165900/

Change-Id: I128fed1380e07a8e540eebd4c4edbdff144caf41
Signed-off-by: Scott <scott@ubuntu.com>