QS tiles string changes
The following tiles all had one thing in common, they all failed to follow AOSP behavior.
They all had the label show 'off' when the tile was in the off state. Normal AOSP behavior
uses the drawable to indicate whether a tile is off or on. For example the flashlight tile
which comes loaded on AOSP doesn't show Flashlight off when the tile is off. You can look at
the other tiles such as WiFi, Airplane, Cast, Bluetooth and DND.

- Location
- Doze notifications
- Smartbar/Fling

- Show Adb network's name in the disabled state. This was showing 'Disabled' only which again
doesn't make a whole lot of sense. The label in the off state should show the tile's name.

Change-Id: I678be421885d7c6647e807ce17a8b356594138f3
Signed-off-by: scotthartbti <scott.hart.bti@gmail.com>