QS BluetoothTile proiritizes connected device UPDATE
When there is a connected device and the
Bluetooth Cacheded Device list is too large
it doesn't guarantee a connected device
will show up at all in the Bluetooth
Quick Settings Tile. This code modifies
the tile to match Wifi witch guarantees
connected devices be at the top of the
Bluetooth Tile list.

Bug: 37438599
Test: Manually pair multiple deivces to create
a large list and change connected states.

Change-Id: I39e35075005d84599aa8b7429e0b4da26bcfab2f